// types/ExtensionFile.type.ts
export enum ExtensionFileType {
    GIF = "gif",
    PDF = "pdf",
    DOC = "doc",
    JPEG = "jpg",
    TXT = "txt",
    EXE = "exe",
    ZIP = "zip",
    XLSX = "xlsx",
    XLSM = "xlsm",
    XLS = "xls",
    XLM = "xlm",
    WPD = "wpd",
    RAR = "rar",
    PNG = "png",
    WBK = "wbk",
    MP3 = "mp3",
    MP4 = "mp4",
    WMV = "wmv",
    JAR = "jar",
    ISO = "iso",
    html = "html",
    DOCX = "docx",
    CSV = "csv",
    TMP = "tmp",
    PSD = "psd",
    RTF = "rtf",
    SWF = "swf",
    JSON = "json",
  }
  